Add SpringMVC and its separate SecurityFilterChain.
This is the beginning of integrating a templating engine to make this a full-stack app. It uses **Thymeleaf**. Modified the SecurityConfig to have separate SecurityFilterChains for the REST API and the web endpoints. The REST endpoints would use the existing JWT-based Access/Refresh tokens flow. The web endpoints would use the default FormLogin flows of Spring Security.

Includes placeholder HTML files for "/" "/login" and "/dashboard" endpoints. Which will be properly implemented and styled in the future.